1.0.0 Version
```Refactor SettingActivity to use mutableStateOf for managing the settings list and perform actions based on user selections. Add a link to AboutActivity. By replacing the static settings list with mutableStateOf and handling user click events in SplashSettingScreen, the refactor improves the flexibility and maintainability of the code. The newly added AboutActivity link allows users to access the About page.
```

```重构SettingActivity以使用mutableStateOf进行设置列表管理，并根据用户选择项执行操作。添加AboutActivity的链接。通过使用mutableStateOf替换静态设置列表，并在SplashSettingScreen中处理用户点击事件，重构提高了代码的灵活性和可维护性。新增的AboutActivity链接允许用户访问关于页面。
```

+++ app/src/main/java/com/gohj99/telewatch/SettingActivity.kt
@@ -1,3 +1,11 @@
+/*
+ * Copyright (c) 2024 gohj99. Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+ * Morbi non lorem porttitor neque feugiat blandit. Ut vitae ipsum eget quam lacinia accumsan.
+ * Etiam sed turpis ac ipsum condimentum fringilla. Maecenas magna.
+ * Proin dapibus sapien vel ante. Aliquam erat volutpat. Pellentesque sagittis ligula eget metus.
+ * Vestibulum commodo. Ut rhoncus gravida arcu.
+ */
+
 package com.gohj99.telewatch
 
 import android.content.Context
@@ -9,30 +17,55 @@ import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.gohj99.telewatch.ui.SettingItem
-import com.gohj99.telewatch.ui.SplashSettingScreen
+import androidx.compose.runtime.mutableStateOf
+import com.gohj99.telewatch.ui.setting.SplashSettingScreen
 import com.gohj99.telewatch.ui.theme.TelewatchTheme
 import java.io.File
 
 class SettingActivity : ComponentActivity() {
+    private var settingsList = mutableStateOf(listOf<String>())
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         enableEdgeToEdge()
 
-        val settingsList = listOf(
-            SettingItem.NestedSetting(getString(R.string.Clearing_cache), onClick = {
-                cacheDir.deleteRecursively()
-                Toast.makeText(this, getString(R.string.Successful), Toast.LENGTH_SHORT).show()
-            }),
-            SettingItem.NestedSetting(getString(R.string.Reset_libtd), onClick = { reset_libtd() }),
-            SettingItem.NestedSetting(getString(R.string.Reset_self), onClick = { reset_self() })
+        settingsList.value = listOf(
+            getString(R.string.Clearing_cache),
+            getString(R.string.Reset_libtd),
+            getString(R.string.Reset_self),
+            getString(R.string.About)
         )
 
         setContent {
             TelewatchTheme {
                 SplashSettingScreen(
-                    settings = settingsList
+                    settings = settingsList,
+                    callback = { item ->
+                        when (item) {
+                            getString(R.string.Clearing_cache) -> {
+                                cacheDir.deleteRecursively()
+                                Toast.makeText(
+                                    this,
+                                    getString(R.string.Successful),
+                                    Toast.LENGTH_SHORT
+                                ).show()
+                            }
+
+                            getString(R.string.Reset_libtd) -> {
+                                reset_libtd()
+                            }
+
+                            getString(R.string.Reset_self) -> {
+                                reset_self()
+                            }
+
+                            getString(R.string.About) -> {
+                                val intent = Intent(this, AboutActivity::class.java)
+                                startActivity(intent)
+                            }
+                        }
+                    }
                 )
             }
         }
